# docs/swagger.yaml
basePath: /
definitions:
  auth.CallbackResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      redirectURL:
        type: string
    type: object
  protocol.HTTPResponse:
    properties:
      data: {}
      error:
        type: string
    type: object
  protocol.RefreshTokenBody:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  protocol.RefreshTokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  protocol.UpdateUserBody:
    properties:
      userName:
        type: string
    required:
    - userName
    type: object
  user.CurUserDTO:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      lastLogin:
        type: string
      name:
        type: string
      permission:
        type: string
      userID:
        type: integer
    type: object
  user.CurUserInfoResponse:
    properties:
      user:
        $ref: '#/definitions/user.CurUserDTO'
    type: object
  user.UpdateUserInfoResponse:
    properties:
      success:
        type: boolean
    type: object
  user.UserDTO:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      lastLogin:
        type: string
      name:
        type: string
      userID:
        type: integer
    type: object
  user.UserInfoResponse:
    properties:
      user:
        $ref: '#/definitions/user.UserDTO'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Go Backend Tmpl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Go Backend Tmpl
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Health check endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: string
                error:
                  type: object
              type: object
      summary: Ping
      tags:
      - health
  /v1/oauth2/{provider}/callback:
    get:
      consumes:
      - application/json
      description: 处理OAuth2授权回调
      parameters:
      - description: OAuth2提供商 (github/google)
        in: path
        name: provider
        required: true
        type: string
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态码
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.CallbackResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      summary: OAuth2回调
      tags:
      - oauth2
  /v1/oauth2/{provider}/login:
    get:
      consumes:
      - application/json
      description: 获取OAuth2授权URL
      parameters:
      - description: OAuth2提供商 (github/google)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      summary: OAuth2登录
      tags:
      - oauth2
  /v1/token/refresh:
    post:
      consumes:
      - application/json
      description: 使用refresh token获取新的access token
      parameters:
      - description: 刷新令牌请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/protocol.RefreshTokenBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/protocol.RefreshTokenResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      summary: 刷新访问令牌
      tags:
      - token
  /v1/user:
    patch:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 更新用户信息请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/protocol.UpdateUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.UpdateUserInfoResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - user
  /v1/user/{userID}:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 用户ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.UserInfoResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - user
  /v1/user/current:
    get:
      consumes:
      - application/json
      description: 获取当前用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.CurUserInfoResponse'
                error:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"